Move code from torch/__init__.py to torch/_overrides.py
Remove utility code that we can simply import from NumPy for now.
